adds company scraper to server
