+++ server/routes.py
@@ -5,12 +5,19 @@ from server.controllers.people import PeopleAPI, PersonAPI
 from server.controllers.getjobs import GetJobsAPI
 from server.controllers.users import UsersAPI
 from server.controllers.sessions import SessionsAPI
-from server.controllers.web_scraper import WebScraperAPI
+from server.controllers.people_scraper import PeopleScraperAPI
+from server.controllers.company_scraper import CompanyScraperAPI
 
 api.add_resource(
-    WebScraperAPI,
-    '/api/webscraper',
-    endpoint='webscraper'
+    CompanyScraperAPI,
+    '/api/companyscraper',
+    endpoint='companyscraper'
+)
+
+api.add_resource(
+    PeopleScraperAPI,
+    '/api/peoplescraper',
+    endpoint='peoplescraper'
 )
 
 api.add_resource(
